% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occlusion.R
\name{extractOcclusionRegions}
\alias{extractOcclusionRegions}
\title{Extracts occlusion regions from input, input occlusion and inpainted output images}
\usage{
extractOcclusionRegions(object)
}
\arguments{
\item{object}{Inpaint class object}
}
\value{
List with occlusion class, containing three occluded regions
}
\description{
Takes inpaint object with input, input occlusion and inpainted output images. Creates a new occlusion class and returns the object.
}
