% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacakge.R
\docType{package}
\name{image-Inpainting}
\alias{image-Inpainting}
\alias{image-inpainting-package}
\alias{image-Inpainting-package}
\title{Perform Image Inpainting using advanced algorithm built on C/C++ languages.}
\description{
An implementation of the "Image Inpainting" algorithm explained at <http://www.ipol.im/pub/art/2017/189/>.
}
