% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageInpaint.R
\name{imageInpaint}
\alias{imageInpaint}
\title{Apply image inpainting algorithm to the input images.}
\usage{
imageInpaint(fileIn, fileInOcc, fileOut, patchSizeX = 7L,
  patchSizeY = 7L, nLevels = -1, useFeatures = 1, verboseMode = 0)
}
\arguments{
\item{fileIn}{Input image path.}

\item{fileInOcc}{Image path with the occlusion.}

\item{fileOut}{Output image path after application of image inpainting.}

\item{patchSizeX}{Size of X patch (3L-13L). 7L is most suitable for images between 512x512 to 800x800 pixels}

\item{patchSizeY}{Size of Y patch (3L-13L). 7L is most suitable for images between 512x512 to 800x800 pixels}

\item{nLevels}{Number of pyramidlevels (0-15). Defaults to automatic selection by the algorithm.}

\item{useFeatures}{Use features. If enabled, the algorithm also matches the region's texture while inpainting}

\item{verboseMode}{Enables or disables the verbose mode.}
}
\value{
as list containing time taken and path of the output image.
}
\description{
An implementation of the "Image Inpainting" algorithm explained at <http://www.ipol.im/pub/art/2017/189/>.
}
