% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageInpaint.R
\name{plot.inpaint}
\alias{plot.inpaint}
\title{Plots three inpaint images}
\usage{
\method{plot}{inpaint}(x, ...)
}
\arguments{
\item{x}{Inpaint list object}

\item{...}{Base plot arguments}
}
\description{
Plots input, occlusion input and inpainted output images
}
