% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occlusion.R
\name{plot.occlusion}
\alias{plot.occlusion}
\title{Plots three occlusion regions}
\usage{
\method{plot}{occlusion}(x, ...)
}
\arguments{
\item{x}{Occlusion list object}

\item{...}{Base plot arguments}
}
\description{
Takes an occlusion class object and plots occlutions regions before and after inpainting.
}
