% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageInpaint.R
\name{print.inpaint}
\alias{print.inpaint}
\title{Prints the inpaint object}
\usage{
\method{print}{inpaint}(x, ...)
}
\arguments{
\item{x}{Inpaint object}

\item{...}{Base print arguments}
}
\description{
Prints the inpaint object
}
