% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageInpaint.R
\name{summary.inpaint}
\alias{summary.inpaint}
\title{Summarizes inpaint object}
\usage{
\method{summary}{inpaint}(object, ...)
}
\arguments{
\item{object}{Inpaint list object}

\item{...}{Base summary arguments}
}
\description{
Provides summary of inpaint object in a tabular form.
}
