% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageInpaint.R
\name{imageInpaint}
\alias{imageInpaint}
\title{Apply image inpainting algorithm to the input images.}
\usage{
imageInpaint(fileIn, fileInOcc, fileOut, patchSizeX = 7L,
  patchSizeY = 7L, nLevels = -1, useFeatures = 1, verboseMode = 0)
}
\arguments{
\item{fileIn}{Input image path.}

\item{fileOut}{Output image path after application of image inpainting.}

\item{patchSizeX}{Size of X patch.}

\item{patchSizeY}{Size of Y patch.}

\item{nLevels}{Number of levels.}

\item{useFeatures}{Use features.}

\item{verboseMode}{Verbose mode.}

\item{fileOcc}{Image path with the occlusion.}
}
\value{
as list containing time taken and path of the output image.
}
\description{
An implementation of the "Image Inpainting" algorithm explained at <http://www.ipol.im/pub/art/2017/189/>.
}
